Refactored ONNX example: Split into main.cpp, RF_DETR_ONNX.h, and RF_DETR_ONNX.cpp for better structure and maintainability
